fix: split arm/authorize/catch skeleton into DirectToolInvoker.Arming.cs
Applies /code-review's findings on fix/494-unify-arm-catch-skeleton: the
shared skeleton and its supporting pieces (RunArmedCoreAsync, ArmGovernance,
LogTrace, Refused, the two log templates) move into their own partial file,
following this class's own file-length convention. TimedOut/Faulted split
out of RunArmedCoreAsync to keep it under the function-length limit.

Also computes each surface's effective timeout once and threads it through
by parameter instead of re-deriving it inside AuthorizeAndRunAsync/
AuthorizeAndRunMcpAsync, and switches both RunArmedCoreAsync call sites to
named arguments.

Follow-up filed for the remaining AuthorizeAndRunAsync/AuthorizeAndRunMcpAsync
duplication (#547) — one level deeper than this PR's scope.
